Working on model for brewery.

+++ simulation/src/main/java/systems/brewery/model/instructions/AddIngredient.java
@@ -8,8 +8,14 @@ import systems.brewery.model.Ingredient;
 @AllArgsConstructor(staticName = "apply")
 public class AddIngredient implements Instruction {
 
+    /**
+     * The ingredient which should be added.
+     */
     Ingredient ingredient;
 
-    int amount;
+    /**
+     * The amount of the ingredient (the unit is part of the ingredient definition).
+     */
+    double amount;
 
 }
